+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 
 * Drop support for python 2.6 and python 3.3
 * Handle changing connections during redistribute ready
+* Add create topic and create channel to LookupdClient
 * [DEPRECATED] :class:`gnsq.Nsqd` is deprectated. Use
   :class:`gnsq.NsqdTCPClient` or :class:`gnsq.NsqdHTTPClient` instead.
 * [DEPRECATED] :class:`gnsq.Lookupd` is deprectated. Use